Pinecone_Product_Similarity/requirements.txt
pinecone>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
